requirements.txt
asgiref==3.5.2
certifi==2021.5.30
Django==3.2.14
django-extensions==3.2.1
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2==2.9.4
pytz==2022.2.1
recursos-humanos==0.1
sqlparse==0.4.2
